+++ lib/urlwatch/reporters.py
@@ -404,6 +404,23 @@ class EMailReporter(TextReporter):
         mailer.send(msg)
 
 
+class IFTTTReport(TextReporter):
+    """Send summary via IFTTT"""
+
+    __kind__ = 'ifttt'
+
+    def submit(self):
+        webhook_url = 'https://maker.ifttt.com/trigger/{event}/with/key/{key}'.format(**self.config)
+        for job_state in self.report.get_filtered_job_states(self.job_states):
+            pretty_name = job_state.job.pretty_name()
+            location = job_state.job.get_location()
+            result = requests.post(webhook_url, json={
+                'value1': job_state.verb,
+                'value2': pretty_name,
+                'value3': location,
+            })
+
+
 class WebServiceReporter(TextReporter):
     MAX_LENGTH = 1024